pyproject.toml
[tool.poetry]
name = "npc-engine"
version = "0.1.0"
description = "Intelligent NPC backend framework powered by Google ADK for game developers"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "npc_engine"}]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^1.0.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
litellm = "^1.0.0"
asyncio = "^3.4.3"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.adk]
name = "npc_engine"
description = "Intelligent NPC backend framework" 